+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Methods of actors can now participate in the actor protocol directly via
   Celluloid#receive
 * Configure custom mailbox types using Celluloid.use_mailbox
+* Define a custom finalizer for an actor by defining MyActor#finalize
 * Actor.call and Actor.async API for making direct calls to mailboxes
 * Fix bugs in Celluloid::Supervisors which would crash on startup if the actor
   they're supervising also crashes on startup